backend/requirements.txt
# ========================================
# 掘进工作面规程智能生成平台 — Python 依赖
# ========================================

# --- Web 框架 ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# --- 数据库 ---
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.2.4

# --- 数据校验 ---
pydantic>=2.5.0
pydantic-settings>=2.1.0

# --- 认证 ---
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# --- Redis ---
redis[hiredis]>=5.0.0

# --- 异步任务 ---
celery>=5.3.0

# --- 文档生成 ---
python-docx>=1.1.0
Jinja2>=3.1.0

# --- AI / LLM ---
langchain>=0.1.0
langfuse>=2.0.0

# --- 测试 ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# --- 工具 ---
python-dotenv>=1.0.0
